before remove onKeyPress
