--- src/components/atoms/FirstDest.js
@@ -1,9 +1,0 @@
-function FirstDest(){
-    return(
-        <>
-            <img className="w-theme-width h-theme-height object-cover rounded-xl" src={imgSrc} alt="default"></img>
-
-        </>
-    );
-}
-export default FirstDest;
