+++ src/components/organisms/FooterT.js
@@ -1,6 +1,13 @@
 function FooterT(){
     return(
-        
-    )
+        <>
+      <div className="w-full h-180 bg-real-blue flex justify-center items-center">
+        <strong className="text-white justify-center items-center">이 여행이 마음에 든다면?</strong>
+        <button className="text-white top-0">
+            더알아보기
+        </button>
+      </div>
+        </>
+    );
 }
 export default FooterT;
