+++ src/components/template/ShowResult.js
@@ -1,0 +1,22 @@
+import FirstBox from "../organisms/FirstBox";
+import SecondBox from "../organisms/SecondBox";
+import ThirdBox from "../organisms/ThirdBox";
+import FooterT from "../organisms/FooterT";
+import HeaderT from "../organisms/HeaderT";
+import Footer from "../organisms/Footer";
+function ShowResult(){
+    return(
+        <div className="flex flex-col">
+            <HeaderT/>
+            <FirstBox/>
+            <SecondBox />
+            <ThirdBox/>
+            <FooterT/>
+            <Footer/>
+        </div>
+    )
+    
+
+}
+
+export default ShowResult;
