--- src/pages/DayonePage.js
@@ -1,11 +1,0 @@
-import HeaderT from "../components/organisms/HeaderT";
-import ChoiceButn from "../components/organisms/ChoiceButn";
-function DayonePage(){
-    return(
-        <>
-        <HeaderT/>
-        <ChoiceButn/>
-        </>
-    );
-}
-export default DayonePage;
