+++ src/pages/ResultPage.js
@@ -1,10 +1,11 @@
-import ChoiceButn from "../components/organisms/ChoiceButn";
 import HeaderT from "../components/organisms/HeaderT";
+import ShowResult from "../components/template/ShowResult";
+import Footer from "../components/organisms/Footer";
+import FooterT from "../components/organisms/FooterT";
 function ResultPage(){
     return(
             <>
-            <HeaderT/>
-            <ChoiceButn/>
+            <ShowResult/>
             </>
     )
