+++ src/pages/SignUpPage.js
@@ -1,5 +1,5 @@
-import Header from "../components/organisms/header";
-import Footer from "../components/organisms/footer";
+import Header from "../components/organisms/Header";
+import Footer from "../components/organisms/Footer";
 import SignUp from '../components/templates/SignUp';
 
 function SignUpPage() {
